Map fix (return from map)
System no longer clears the current feature, nor tries to reset to
previous feature, when exiting the map.  This is b/c with the new
scanner, a map clue is almost exactly like the other clues (with the
appropriate icon displayed in the scanner center) so we can treat it
like any control, for the most part.