Map tags
Map tags are now added at the continent and region levels, in addition
to the country level which was in the last checkin.  Code to add the
tags is now in map.js.
File dimensions for tags are now fixed at 128x128.
Lat checkin before attempting to move CFS stores to Amazon S3.