hack and debrief changes
Renamed hack.startNewFromMenu() to hack.startNext().  This function no
longer inits the display, since that happens in hack.startNew(), which
is called before every hack.
Debrief calls hack.debrief.set() now in debrief.startNew(), so that the
image has plenty of time to preload.
Debrief.set() no longer plays the language file, if appropriate.  This
is now done in deb.template.rendered and in the next / prev deb. nav
buttons.