Editor and S3: first pass
First pass at integrating the database editor with the CFS / S3
collections.  The Image collection is now set to properly display the
S3 stored images, and also to insert / update to that collection.  (The
file objects themselves can’t be updated in the collection, so we do a
delete / insert sequence to make this happen.)