S3 videos
The video collection has been split into 2 — the ghV collection is the
master collection and the f property in these records has this format
for non-YT vids:
s3@ID
where ID is the _id of the record in ghPublicVideo that links to the
file that was uploaded to S3.  Otherwise YT files remain unchanged; the
f property is the YT ID of the file.
The game and the editor have been fixed to accept these changes, for
displaying the videos, but not yet for editing the video records.