Editor: video recs, edit this country menu item
Editing video records had a bug caused by the editor.youTubeLoaded
property, which has been deleted in favor of the global property,
youTubeLoaded.

Edit this country — on the menu — was fixed to get the country from the
global hack.  Still need to either disable/hide this option if there is
no country, or else use the user’s home country.