More db changes, plus Slingshot
All content collections are single, standard collections now.
(ghAvatar still has a CFS collection, solely for the purpose of
uploading the initially generated avatar from the server.)  The content
collections all use the u field for their URL, which either points to
an S3 object or to something else (like a streaming sound site).  The
exception is YT videos, which have their YT ID in the u field.

This new db arrangement works for reading the files, but not yet for
creating, updating or deleting.

Slingshot code added back in.  SS directives and uploaders exist for
all the content collections.