Tags normalized, editor fix
Normalized the data for ghTag.  This collection is now just the dt and
cc fields (deb. type and country code).  The tag objects that the user
keeps in profile have been reduced to cc, dt and long. and lat.
All tags now require a corresponding record in ghText that match on cc
and dt; this is where the text for the tag is.  The various functions
that need the tag URL retrieve it from ghTag, using the same fields.
Added a modal requester to get the debrief code when a tag is added.
Tags now show up per user, per country on the regular gameplay map.
A bug preventing editing text records in the db was fixed.