Clickable map images, new music
Map images are now clickable and the result is the same as clicking on
the country, or will be once the zoom vars are written to the country
recs.
Fixed the ghImageCtl object so that works correctly with changes to the
collection (redacted maps) in instantMode.
instantMode now also properly disables the scan button when no more
data is available.
Added a new set of music tracks.