HTTP, retry new avatar URL
Added HTTP package for an attempt at overcoming an issue with newly
created avatars.  The problem is that the URLS for newly-created avatar
files, stored in S3 by CFS, are not immediately available.  A 2 second
delay before trying worked fine until yesterday (a couple of attempts
at new user creation failed to get a working avatar link), so I added
an HTTP get request on the URL, before setting that value in the
database.  But the test always fails, so I’ve changed the timeOut to 3
seconds, and I’m hoping to resolve the HTTP issue.

Also, fixed a bug preventing “list all users …” from working.  Added a
new global in constants:  avatarPrefix.