Avatar URL checking
The CFS package is used to upload a newly-created avatar (for a newly
created user) to S3.  This is dicey b/c:
1) the file object returned by the CFS insert method does not have a
name or URL available for immediate use and …
2) the URL, once it is available, is not necessarily valid immediately
Now using the HTTP package to test the URL with a 3 second timeout,
once the name() function returns a value.  If the URL fails, then we
re-test it (up to 5 times).