Delete game.user.profile.av
The avatar field (av) in the game.user.profile has been eliminated — at
least any explicit use of it in the game has been.  Because of all the
various operations that happen on the server to create and upload the
initial avatar file, it has proven difficult to get the final URL back
from the server at creation time.  Given this, we are just using
Meteor.user().profile.av throughout the code now, which works fine.