Oceania top 7, lesson enhancements
Added a mission for the Top 7 Countries in Oceania.
Properly integrated lesson-capsule editing into in-game-editor.js
Enhanced “click on a country name to learn” with text to indicate
region and a pulse on the country’s opacity.
Added a new region field for region label color in learn mode:  rll_co