Browsing while learning
User can now browse a country by clicking on the learn capsule.  Added
a Return to Lesson button on main (browse) screen.
Lesson buttons are now proportionally sized, so as not to overflow the
available space.
resumeLesson() functionality added, to allow user to return to where
they were in a lesson, should they leave that screen prematurely (to
browse, say).
Browsing not permitted during quiz.  Added a trap and an informational
message for this.
Completion tick marks added to lesson buttons.