Help videos
This is the first phase of replacing the previous intro video with a
new help screen that features 4 videos that the user can watch at their
leisure.

Last checkin before attempting to rewrite the code in
onYouTubeIframeAPIReady() which is convoluted.

This version wired to go straight to Instructions screen and may be
broken in some other ways too.