New video -- images and web
Images and web are loading correctly as features — the new video scheme
has now entailed a major rewrite of the feature object.

Simplified operations include:
No preview of the feature in the center — feature automatically loads
after scan.
No tracking of lastName in feature — map object  will be rewritten to
no longer be a “feature”, but an independent functionality.
feature.name will likely switch to a regular var, not a ReactiveVar.