New video scheme -- editor
Editor now functions with new scheme.
Fixed some navigation issues to / from the editor.
Fixed a bug in control initialization that was getting the fullCount
property wrong.
Fixed an issue that was causing the music to play over videos in hack
mode.
Restored ‘jump to browse country’ feature on agents screen (click the
flag).
Made the scanner fade out at end of scan.