Various final fixes
A few more fixes before folding this branch back into main.
The user’s pic on the congrats screen now goes to user’s bio page.
db.getRandomRecForUser now checks to make sure that the country is
still in the database before  passing it back to hack object.  (Some
countries were deleted after the user missions were created for certain
older users, like Svalbard.)
Return button in browser now displays a user pic when returning to home
screen and a geohacker logo when returning to mission and a globe when
returning to map.