New navigation mode
Changed the “navigate countries” mode (editing feature) to
automatically be active if user is status Admin and user is currently
browsing or editing a country in the db.
Made the window not scroll horizontally either on certain screens like
browse.
Added “artist” to data types for image.
Fixed a browse issue that was causing a meme to show twice in a row on
same side.