More meme integration
Memes (as clues) now fade-in and dim appropriately.
Memes work with nav buttons on hacker screen.
Memes in Explore can be clicked on to view normally (nor rotated) and
can then be clicked on to return to Explore.
Changed method of hiding / showing meme on hacker screen — not a
reactive change anymore.
Deleted feature.clear(), but may need something like it for “continue
mission” scenario.
Added a check to see if new field “rl_co” exists in region recs for
region label color.