Reload browsing data
Added a menu item to reload the current browse of a country.  This
makes testing changes to the country’s data easier.  A function to do
this automatically when going from editor to browse does not appear to
be working, but the function called by the menu item does work.