Less reactivity, more query (story mode)
To better control the visibility of the story objects, we switched to
jquery for controlling the objects.  (Basic layout, i.e., the number of
identically structured story objects on the DOM is still a reactive
feature).

Inanimate objects are now classed as tokens (token.js)  These objects
can have an array of content images, should the main object need to be
a frame or a screen, that other images are displayed on.