Story objects, storyA adds
Moved story specific content into it’s own folder.
Reworked the scene progression, so that the story object can decide
which scene to play based on the location of the scene and the data
flags.
Added ability to jump into a different chat sequence in medias res.
Added ability to play a scene as a helper response to a chat choice.
Added more to storyA.