Stories -- transition to db
Story data is migrating to the database.  Story now subscribes to the
relevant data when inited.  The only data item being used at this point
is the user records for any agents in the story who are in the user db.

Next:  replace the hard-coded entity definitions with entities created
in real-time from the db.