Story chat editor -- create new chat
New record button on chat data screen now asks the user for a name for
the chat, and then creates the first two records needed for the chat.
User responses now have move buttons to move the response up and down
in the array.
Changed some flag names in preparation for running the story with flag
values set in the editor.  (has_[token] and gave_[token] will update
inventory).