Data / Visual editors integration 2
Removed the scene table (ghScene) from the code b/c it was serving no
purpose.  ghCue does all the work here.  Internally, most of the code
refers to “cue” but to the user, a cue is a scene.  Story.scene is now
the cue name.

Made the Play button on visual editor check and go to the current story
location, if a story is already loaded and we have not yet picked a new
one.
Scene (cue) and Chat buttons in VED now load the expected data if you
use them during story play.