SMED fixes
SMED was not testing the dest names to see if they were real and so was
creating “exit” records.  Fixed.
createDestinationName() in SMED was not using the destinationList
correctly to test for duped names.  Fixed.
Added a function to delete entire chat — needs a confirmation requester
and testing.
Outlined a scrubData function.