Entity edits and commands
Entity edits to size and pos now respect the current transform values
of the entity.
Added skewX and skewY properties to entities.
VED now displays coordinates to the screen as an object; the object
format that entity.add() now expects as an optional argument, so size
and pos can be set in add() without requiring an additional change()
command.