Add transform values from VED to cue
The cog button on the cue records in the editor launches the ‘Add scale
/ position’ values modal box.  The checkboxes on this box allow you to
select which value to copy over to the current command.  The ‘Add’
buttons adds a object in parentheses to the end of the current command
with the selected name:value pairs.

Changed moveTo command to animate.  Can animate scale, position and
opacity, with an optional duration value.

New prop in VED, recordIDToEdit.  When an entity is selected in VED,
it’s ID is stored in sed.recordID  So we now have this additional
property in VED to hold the recordID of the selected record when we
call up the ‘add transform modal’.