Story exercise 2
Exercise objects improved.  Stories must now call em.build() before
starting an exercise.  If an ID is passed to build(), it is assumed to
signify a complete set of items that the Exercise object can create
just from the ID (“where are all the continents”, e.g.)

If no ID is passed to build(), then the code in story will have to call
em.add() with an array of objects, each of which will become an item.

“inWhichContinent” is  a new exercise item that was added.

Code for the Map button on story screen now resets the browseMap before
opening.

The routine for starting an exercise is documented in the story object
comments.  Basically em.build() switches the template to “exercise” and
that template’s rendered event calls em.go()