Navigation Ready check
Story object now waits 2 seconds after launching the scene to enable
the navigation (base and map) buttons.  The map button will not work if
not enabled.  This is to prevent the story from drawing entities on top
of the map (if you click too soon, the cutScene may still have commands
in the queue.)