Story changes, font size editing on browser
Countries now have one destination each.  Defaults to capital, but can
be changed in story.getCityName().

story.playDefaultScene now takes a countryCode parameter, if missing,
then it sets the bg image to be from the capital, otherwise it expects
a bg image in the locations table.

Font size editing restored in new browser.