storyB, custom exercises, story labels
Much work on storyB, which is nearly done.

Custom exercises can now be created using ammap.  In storyB, we use a
custom version of the Asia map that has the islands of Indonesia
decomposed into separate areas, so that we can detect clicks on each.

Stories can now use the new storyLabel object, which allows for the use
of configurable text objects that are similar to entities, though they
do not use CSS transforms.