Story engine tweaks; completion of StoryB
Story B complete, save a few possible tweaks before release.
Fixed / recoded the handling of the story prompt.  Stories can now show
a prompt based on an inventory item addition; saves having to add an
additional scene in some cases.  The default is to automatically hide
the prompt after an inv add, but this can now be over-ridden in the
story script.
Restored display base name as the city name when user is there.
Z-index:  moved tokens in front of chars
Char:  added actLikeToken property.  Set this to true to make a char
act like an inventory item — lets user transport a char somewhere.