News feed
Status line at top of Hacker screen is now a “news” feed.  News is:
—local events (current user’s stats / accomplishments)
—weather reports
—network events (remote user’s accomplishments)
Basic functionality in place, but there remains a need to automatically
prune the db records (maybe keep a max of 100).  But the records can
potentially contain valuable network history, so we will likely write
out “dying” records to a text file.