URLs for images, disabled IG login
In the editor, Images can now take just a URL, instead of a file.
Primarily for IG images.

Disabled Instagram login b/c only “sandbox” users can get in.

Disabled the screening out of “primary” videos, which determined which
videos to put on the lower line of videos.  New design for this screen
has only one line of videos.