Auto mode 4
More changes to make the rows of controls move up the screen properly
and to make the ones behind the featured item get dimmed properly.

Tweaked the size on the featured item display ( to better match the
rows behind it) and hid the featured bg box.  Added the new GH logo.